+++ src/todo-app/todo-app.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, Injectable, HttpStatus } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { TodoDTO } from './models/todo-app.dto';
@@ -13,7 +13,11 @@ export class TodoAppService {
     }
 
     async getTaskById(id: number) {
-        return await this.todoRepository.findOne({where: {id}});
+        const todo = await this.todoRepository.findOne({where: {id}});
+        if (!todo) {
+            throw new HttpException('Not Found', HttpStatus.NOT_FOUND)
+        }
+        return todo;
     }
 
     async addNew(data: TodoDTO) {
@@ -23,12 +27,20 @@ export class TodoAppService {
     }
 
     async updateTask(id: number, data: TodoDTO) {
+        const todo = await this.todoRepository.findOne({where: {id}});
+        if (!todo) {
+            throw new HttpException('Not Found', HttpStatus.NOT_FOUND)
+        }
         await this.todoRepository.update({id}, data);
-        return await this.todoRepository.findOne({id});
+        return todo;
     }
 
     async deleteTask(id: number) {
+        const todo = await this.todoRepository.findOne({where: {id}});
+        if (!todo) {
+            throw new HttpException('Not Found', HttpStatus.NOT_FOUND)
+        }
         await this.todoRepository.delete({id});
-        return {deleted: true};
+        return todo;
     }
 }
